Finished library refactoring.
Completed refactoring all code so new Qt ACE now works as a library.
Also finished dataobject class.

+++ qt/src/abstractanalyticfactory.h
@@ -16,9 +16,9 @@ public:
    ACE_MAKE_NO_COPY_OR_MOVE(EAbstractAnalyticFactory);
    static EAbstractAnalyticFactory& getInstance();
    static void setInstance(std::unique_ptr<EAbstractAnalyticFactory>&& factory);
-   virtual int getCount() = 0;
-   virtual QString getName(int type) = 0;
-   virtual std::unique_ptr<EAbstractAnalytic> makeAnalytic(int type) = 0;
+   virtual quint16 getCount() = 0;
+   virtual QString getName(quint16 type) = 0;
+   virtual std::unique_ptr<EAbstractAnalytic> makeAnalytic(quint16 type) = 0;
 private:
    static std::unique_ptr<EAbstractAnalyticFactory> _instance;
 };
